feature1

+++ src/main/java/Investment.java
@@ -2,7 +2,7 @@
 
 public class Investment extends Account {
 
-    public Investment(float initialDeposit, String type){
+    public Investment(double initialDeposit, String type){
         super (initialDeposit,type);
     }
